re-organize code
